Dcoument casttoint.cpp

+++ miki-framework/extra/casttoint.cpp
@@ -1,5 +1,9 @@
 namespace extra {
-    int castToInt(double value) {
+    int castToInt(float value) {
+        /**
+         * @brief This function converts int number to float
+         * @param value Number to convert
+         */
         return static_cast<int>(value);
     }
 }
